refactored code by renaming classes and methods from Zend to Laminas, adding deprecated proxy classes/methods for the old names to be dropped in the next major release
