+++ src/Phpro/SoapClient/CodeGenerator/LaminasCodeFactory/DocBlockGeneratorFactory.php
@@ -1,13 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace Phpro\SoapClient\CodeGenerator\ZendCodeFactory;
+namespace Phpro\SoapClient\CodeGenerator\LaminasCodeFactory;
 
 use Laminas\Code\Generator\DocBlockGenerator;
 
-/**
- * @deprecated Please use LaminasCodeFactory\DocBlockGeneratorFactory instead
- * @see \Phpro\SoapClient\CodeGenerator\LaminasCodeFactory\DocBlockGeneratorFactory
- */
 final class DocBlockGeneratorFactory
 {
     public static function fromArray(array $data): DocBlockGenerator
