+++ src/Phpro/SoapClient/CodeGenerator/Util/Validator.php
@@ -7,17 +7,35 @@ use Laminas\Code\Generator\ClassGenerator;
 
 class Validator
 {
-    public static function zendCodeIsInstalled(): bool
+    public static function laminasCodeIsInstalled(): bool
     {
         return class_exists(ClassGenerator::class);
     }
 
-    public static function commandRequiresZendCode(string $name): bool
+    /**
+     * @deprecated use laminasCodeIsInstalled() instead
+     * @see self::laminasCodeIsInstalled()
+     */
+    public static function zendCodeIsInstalled(): bool
+    {
+        return self::laminasCodeIsInstalled();
+    }
+
+    public static function commandRequiresLaminasCode(string $name): bool
     {
         if ($name === WizardCommand::COMMAND_NAME) {
             return true;
         }
 
         return strpos($name, 'generate') === 0;
+    }
+
+    /**
+     * @deprecated use commandRequiresLaminasCode() instead
+     * @see self::commandRequiresLaminasCode()
+     */
+    public static function commandRequiresZendCode(string $name): bool
+    {
+        return self::commandRequiresLaminasCode($name);
     }
 }
